src/test/javascript/spec/app/entities/doente/doente-update.component.spec.ts
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { of } from 'rxjs';

import { Demografia2TestModule } from '../../../test.module';
import { DoenteUpdateComponent } from 'app/entities/doente/doente-update.component';
import { DoenteService } from 'app/entities/doente/doente.service';
import { Doente } from 'app/shared/model/doente.model';

describe('Component Tests', () => {
  describe('Doente Management Update Component', () => {
    let comp: DoenteUpdateComponent;
    let fixture: ComponentFixture<DoenteUpdateComponent>;
    let service: DoenteService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [Demografia2TestModule],
        declarations: [DoenteUpdateComponent],
        providers: [FormBuilder]
      })
        .overrideTemplate(DoenteUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(DoenteUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(DoenteService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new Doente(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new Doente();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
